split sonarr into more files
